Add CORS support to DistributoreController and update polling logic for user data retrieval

+++ Progetto-Coffee-cApp/src/main/java/it/unipa/wsda/progettocoffeecapp/model/Bevanda.java
@@ -11,5 +11,5 @@ public class Bevanda {
     @Id
     private Integer id_bevanda;
     private String nome;
-    private Integer costo;
+    private Double costo;
 }
